+++ pkg/reconciler/claimbinding/reconciler.go
@@ -32,10 +32,12 @@ import (
 	"github.com/crossplaneio/crossplane-runtime/apis/core/v1alpha1"
 	"github.com/crossplaneio/crossplane-runtime/pkg/logging"
 	"github.com/crossplaneio/crossplane-runtime/pkg/meta"
+	"github.com/crossplaneio/crossplane-runtime/pkg/resource"
 )
 
 const (
 	claimControllerName   = "resourceclaim.crossplane.io"
+	claimFinalizerName    = "finalizer." + claimControllerName
 	claimReconcileTimeout = 1 * time.Minute
 
 	aShortWait = 30 * time.Second
@@ -54,36 +56,18 @@ const (
 
 var log = logging.Logger.WithName("controller")
 
-// A ClaimKind contains the type metadata for a kind of resource claim.
-type ClaimKind schema.GroupVersionKind
-
-// A ClassKind contains the type metadata for a kind of resource class.
-type ClassKind schema.GroupVersionKind
-
-// List returns the list kind associated with a ClassKind.
-func (k ClassKind) List() schema.GroupVersionKind {
-	return schema.GroupVersionKind{
-		Group:   k.Group,
-		Version: k.Version,
-		Kind:    k.Kind + "List",
-	}
-}
-
-// A ManagedKind contains the type metadata for a kind of managed resource.
-type ManagedKind schema.GroupVersionKind
-
 // A ManagedConfigurator configures a resource, typically by converting it to
 // a known type and populating its spec.
 type ManagedConfigurator interface {
-	Configure(ctx context.Context, cm Claim, cs Class, mg Managed) error
+	Configure(ctx context.Context, cm resource.Claim, cs resource.Class, mg resource.Managed) error
 }
 
 // A ManagedConfiguratorFn is a function that satisfies the
 // ManagedConfigurator interface.
-type ManagedConfiguratorFn func(ctx context.Context, cm Claim, cs Class, mg Managed) error
+type ManagedConfiguratorFn func(ctx context.Context, cm resource.Claim, cs resource.Class, mg resource.Managed) error
 
 // Configure the supplied resource using the supplied claim and class.
-func (fn ManagedConfiguratorFn) Configure(ctx context.Context, cm Claim, cs Class, mg Managed) error {
+func (fn ManagedConfiguratorFn) Configure(ctx context.Context, cm resource.Claim, cs resource.Class, mg resource.Managed) error {
 	return fn(ctx, cm, cs, mg)
 }
 
@@ -92,95 +76,78 @@ func (fn ManagedConfiguratorFn) Configure(ctx context.Context, cm Claim, cs Clas
 // responsible for final modifications to the claim and resource, for example
 // ensuring resource, class, claim, and owner references are set.
 type ManagedCreator interface {
-	Create(ctx context.Context, cm Claim, cs Class, mg Managed) error
+	Create(ctx context.Context, cm resource.Claim, cs resource.Class, mg resource.Managed) error
 }
 
 // A ManagedCreatorFn is a function that satisfies the ManagedCreator interface.
-type ManagedCreatorFn func(ctx context.Context, cm Claim, cs Class, mg Managed) error
+type ManagedCreatorFn func(ctx context.Context, cm resource.Claim, cs resource.Class, mg resource.Managed) error
 
 // Create the supplied resource.
-func (fn ManagedCreatorFn) Create(ctx context.Context, cm Claim, cs Class, mg Managed) error {
+func (fn ManagedCreatorFn) Create(ctx context.Context, cm resource.Claim, cs resource.Class, mg resource.Managed) error {
 	return fn(ctx, cm, cs, mg)
-}
-
-// A ManagedConnectionPropagator is responsible for propagating information
-// required to connect to a managed resource (for example the connection secret)
-// from the managed resource to its resource claim.
-type ManagedConnectionPropagator interface {
-	PropagateConnection(ctx context.Context, cm Target, mg Managed) error
-}
-
-// A ManagedConnectionPropagatorFn is a function that satisfies the
-// ManagedConnectionPropagator interface.
-type ManagedConnectionPropagatorFn func(ctx context.Context, cm Target, mg Managed) error
-
-// PropagateConnection information from the supplied managed resource to the
-// supplied resource claim.
-func (fn ManagedConnectionPropagatorFn) PropagateConnection(ctx context.Context, cm Target, mg Managed) error {
-	return fn(ctx, cm, mg)
 }
 
 // A Binder binds a resource claim to a managed resource.
 type Binder interface {
 	// Bind the supplied Claim to the supplied Managed resource.
-	Bind(ctx context.Context, cm Claim, mg Managed) error
+	Bind(ctx context.Context, cm resource.Claim, mg resource.Managed) error
 
 	// Unbind the supplied Claim from the supplied Managed resource.
-	Unbind(ctx context.Context, cm Claim, mg Managed) error
+	Unbind(ctx context.Context, cm resource.Claim, mg resource.Managed) error
 }
 
 // BinderFns satisfy the Binder interface.
 type BinderFns struct {
-	BindFn   func(ctx context.Context, cm Claim, mg Managed) error
-	UnbindFn func(ctx context.Context, cm Claim, mg Managed) error
+	BindFn   func(ctx context.Context, cm resource.Claim, mg resource.Managed) error
+	UnbindFn func(ctx context.Context, cm resource.Claim, mg resource.Managed) error
 }
 
 // Bind the supplied Claim to the supplied Managed resource.
-func (b BinderFns) Bind(ctx context.Context, cm Claim, mg Managed) error {
+func (b BinderFns) Bind(ctx context.Context, cm resource.Claim, mg resource.Managed) error {
 	return b.BindFn(ctx, cm, mg)
 }
 
 // Unbind the supplied Claim from the supplied Managed resource.
-func (b BinderFns) Unbind(ctx context.Context, cm Claim, mg Managed) error {
+func (b BinderFns) Unbind(ctx context.Context, cm resource.Claim, mg resource.Managed) error {
 	return b.UnbindFn(ctx, cm, mg)
 }
 
 // A ClaimFinalizer finalizes the deletion of a resource claim.
 type ClaimFinalizer interface {
 	// AddFinalizer to the supplied Claim.
-	AddFinalizer(ctx context.Context, cm Claim) error
+	AddFinalizer(ctx context.Context, cm resource.Claim) error
 
 	// RemoveFinalizer from the supplied Claim.
-	RemoveFinalizer(ctx context.Context, cm Claim) error
+	RemoveFinalizer(ctx context.Context, cm resource.Claim) error
 }
 
 // A ClaimFinalizerFns satisfy the ClaimFinalizer interface.
 type ClaimFinalizerFns struct {
-	AddFinalizerFn    func(ctx context.Context, cm Claim) error
-	RemoveFinalizerFn func(ctx context.Context, cm Claim) error
+	AddFinalizerFn    func(ctx context.Context, cm resource.Claim) error
+	RemoveFinalizerFn func(ctx context.Context, cm resource.Claim) error
 }
 
 // AddFinalizer to the supplied Claim.
-func (f ClaimFinalizerFns) AddFinalizer(ctx context.Context, mg Claim) error {
-	return f.AddFinalizerFn(ctx, mg)
+func (f ClaimFinalizerFns) AddFinalizer(ctx context.Context, cm resource.Claim) error {
+	return f.AddFinalizerFn(ctx, cm)
 }
 
 // RemoveFinalizer from the supplied Claim.
-func (f ClaimFinalizerFns) RemoveFinalizer(ctx context.Context, mg Claim) error {
-	return f.RemoveFinalizerFn(ctx, mg)
+func (f ClaimFinalizerFns) RemoveFinalizer(ctx context.Context, cm resource.Claim) error {
+	return f.RemoveFinalizerFn(ctx, cm)
 }
 
-// A ClaimReconciler reconciles resource claims by creating exactly one kind of
+// A Reconciler reconciles resource claims by creating exactly one kind of
 // concrete managed resource. Each resource claim kind should create an instance
 // of this controller for each managed resource kind they can bind to, using
 // watch predicates to ensure each controller is responsible for exactly one
 // type of resource class provisioner. Each controller must watch its subset of
 // resource claims and any managed resources they control.
-type ClaimReconciler struct {
+type Reconciler struct {
 	client     client.Client
-	newClaim   func() Claim
-	newClass   func() Class
-	newManaged func() Managed
+	newClaim   func() resource.Claim
+	newClass   func() resource.Class
+	newManaged func() resource.Managed
 
 	// The below structs embed the set of interfaces used to implement the
 	// resource claim reconciler. We do this primarily for readability, so that
@@ -192,8 +159,7 @@ type ClaimReconciler struct {
 type crManaged struct {
 	ManagedConfigurator
 	ManagedCreator
-	ManagedConnectionPropagator
-	ManagedFinalizer
+	resource.ManagedConnectionPropagator
 }
 
 func defaultCRManaged(m manager.Manager) crManaged {
@@ -203,7 +169,7 @@ func defaultCRManaged(m manager.Manager) crManaged {
 			ManagedConfiguratorFn(ConfigureReclaimPolicy),
 		},
 		ManagedCreator:              NewAPIManagedCreator(m.GetClient(), m.GetScheme()),
-		ManagedConnectionPropagator: NewAPIManagedConnectionPropagator(m.GetClient(), m.GetScheme()),
+		ManagedConnectionPropagator: resource.NewAPIManagedConnectionPropagator(m.GetClient(), m.GetScheme()),
 	}
 }
 
@@ -219,66 +185,72 @@ func defaultCRClaim(m manager.Manager) crClaim {
 	}
 }
 
-// A ClaimReconcilerOption configures a ClaimReconciler.
-type ClaimReconcilerOption func(*ClaimReconciler)
+// A ReconcilerOption configures a Reconciler.
+type ReconcilerOption func(*Reconciler)
 
 // WithManagedConfigurators specifies which configurators should be used to
 // configure each managed resource. Configurators will be applied in the order
 // they are specified.
-func WithManagedConfigurators(c ...ManagedConfigurator) ClaimReconcilerOption {
-	return func(r *ClaimReconciler) {
+func WithManagedConfigurators(c ...ManagedConfigurator) ReconcilerOption {
+	return func(r *Reconciler) {
 		r.managed.ManagedConfigurator = ConfiguratorChain(c)
 	}
 }
 
 // WithManagedCreator specifies which ManagedCreator should be used to create
 // managed resources.
-func WithManagedCreator(c ManagedCreator) ClaimReconcilerOption {
-	return func(r *ClaimReconciler) {
+func WithManagedCreator(c ManagedCreator) ReconcilerOption {
+	return func(r *Reconciler) {
 		r.managed.ManagedCreator = c
 	}
 }
 
 // WithManagedConnectionPropagator specifies which ManagedConnectionPropagator
 // should be used to propagate resource connection details to their claim.
-func WithManagedConnectionPropagator(p ManagedConnectionPropagator) ClaimReconcilerOption {
-	return func(r *ClaimReconciler) {
+func WithManagedConnectionPropagator(p resource.ManagedConnectionPropagator) ReconcilerOption {
+	return func(r *Reconciler) {
 		r.managed.ManagedConnectionPropagator = p
 	}
 }
 
 // WithBinder specifies which Binder should be used to bind
 // resources to their claim.
-func WithBinder(b Binder) ClaimReconcilerOption {
-	return func(r *ClaimReconciler) {
+func WithBinder(b Binder) ReconcilerOption {
+	return func(r *Reconciler) {
 		r.claim.Binder = b
 	}
 }
 
 // WithClaimFinalizer specifies which ClaimFinalizer should be used to finalize
 // claims when they are deleted.
-func WithClaimFinalizer(f ClaimFinalizer) ClaimReconcilerOption {
-	return func(r *ClaimReconciler) {
+func WithClaimFinalizer(f ClaimFinalizer) ReconcilerOption {
+	return func(r *Reconciler) {
 		r.claim.ClaimFinalizer = f
 	}
 }
 
-// NewClaimReconciler returns a ClaimReconciler that reconciles resource claims
+// NewReconciler returns a Reconciler that reconciles resource claims
 // of the supplied ClaimKind with resources of the supplied ManagedKind. It
 // panics if asked to reconcile a claim or resource kind that is not registered
-// with the supplied manager's runtime.Scheme. The returned ClaimReconciler will
+// with the supplied manager's runtime.Scheme. The returned Reconciler will
 // apply only the ObjectMetaConfigurator by default; most callers should supply
 // one or more ManagedConfigurators to configure their managed resources.
-func NewClaimReconciler(m manager.Manager, of ClaimKind, using ClassKind, with ManagedKind, o ...ClaimReconcilerOption) *ClaimReconciler {
-	nc := func() Claim { return MustCreateObject(schema.GroupVersionKind(of), m.GetScheme()).(Claim) }
-	ns := func() Class { return MustCreateObject(schema.GroupVersionKind(using), m.GetScheme()).(Class) }
-	nr := func() Managed { return MustCreateObject(schema.GroupVersionKind(with), m.GetScheme()).(Managed) }
+func NewReconciler(m manager.Manager, of resource.ClaimKind, using resource.ClassKind, with resource.ManagedKind, o ...ReconcilerOption) *Reconciler {
+	nc := func() resource.Claim {
+		return resource.MustCreateObject(schema.GroupVersionKind(of), m.GetScheme()).(resource.Claim)
+	}
+	ns := func() resource.Class {
+		return resource.MustCreateObject(schema.GroupVersionKind(using), m.GetScheme()).(resource.Class)
+	}
+	nr := func() resource.Managed {
+		return resource.MustCreateObject(schema.GroupVersionKind(with), m.GetScheme()).(resource.Managed)
+	}
 
 	// Panic early if we've been asked to reconcile a claim or resource kind
 	// that has not been registered with our controller manager's scheme.
 	_, _, _ = nc(), ns(), nr()
 
-	r := &ClaimReconciler{
+	r := &Reconciler{
 		client:     m.GetClient(),
 		newClaim:   nc,
 		newClass:   ns,
@@ -295,7 +267,7 @@ func NewClaimReconciler(m manager.Manager, of ClaimKind, using ClassKind, with M
 }
 
 // Reconcile a resource claim with a concrete managed resource.
-func (r *ClaimReconciler) Reconcile(req reconcile.Request) (reconcile.Result, error) { // nolint:gocyclo
+func (r *Reconciler) Reconcile(req reconcile.Request) (reconcile.Result, error) { // nolint:gocyclo
 	// NOTE(negz): This method is well over our cyclomatic complexity goal.
 	// Be wary of adding additional complexity.
 
@@ -308,7 +280,7 @@ func (r *ClaimReconciler) Reconcile(req reconcile.Request) (reconcile.Result, er
 	if err := r.client.Get(ctx, req.NamespacedName, claim); err != nil {
 		// There's no need to requeue if we no longer exist. Otherwise we'll be
 		// requeued implicitly because we return an error.
-		return reconcile.Result{}, errors.Wrap(IgnoreNotFound(err), errGetClaim)
+		return reconcile.Result{}, errors.Wrap(resource.IgnoreNotFound(err), errGetClaim)
 	}
 
 	managed := r.newManaged()
@@ -395,7 +367,7 @@ func (r *ClaimReconciler) Reconcile(req reconcile.Request) (reconcile.Result, er
 		}
 	}
 
-	if !IsBindable(managed) && !IsBound(managed) {
+	if !resource.IsBindable(managed) && !resource.IsBound(managed) {
 		if managed.GetClaimReference() == nil {
 			// We're waiting to bind to a statically provisioned managed
 			// resource. We must requeue because our EnqueueRequestForClaim
@@ -414,7 +386,7 @@ func (r *ClaimReconciler) Reconcile(req reconcile.Request) (reconcile.Result, er
 		return reconcile.Result{Requeue: false}, errors.Wrap(r.client.Status().Update(ctx, claim), errUpdateClaimStatus)
 	}
 
-	if IsBindable(managed) {
+	if resource.IsBindable(managed) {
 		if err := r.managed.PropagateConnection(ctx, claim, managed); err != nil {
 			// If we didn't hit this error last time we'll be requeued implicitly
 			// due to the status update. Otherwise we want to retry after a brief
