+++ pkg/reconciler/claimdefaulting/reconciler.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resource
+package claimdefaulting
 
 import (
 	"context"
@@ -28,19 +28,28 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
+	"github.com/crossplaneio/crossplane-runtime/apis/core/v1alpha1"
 	"github.com/crossplaneio/crossplane-runtime/pkg/logging"
 	"github.com/crossplaneio/crossplane-runtime/pkg/meta"
+	"github.com/crossplaneio/crossplane-runtime/pkg/resource"
 )
 
 const (
-	claimSchedulingControllerName       = "resourceclaimscheduler.crossplane.io"
-	claimSchedulingReconcileTimeout     = 1 * time.Minute
-	claimSchedulingReconcileMaxJitterMs = 1500
+	claimDefaultingControllerName       = "resourceclaimdefaulter.crossplane.io"
+	claimDefaultingReconcileTimeout     = 1 * time.Minute
+	claimDefaultingReconcileMaxJitterMs = 1500
+
+	aShortWait = 30 * time.Second
 )
 
+// Error strings.
 const (
+	errGetClaim    = "cannot get resource claim"
+	errUpdateClaim = "cannot update resource claim"
 	errListClasses = "cannot list resource classes"
 )
+
+var log = logging.Logger.WithName("controller")
 
 // A Jitterer sleeps for a random amount of time in order to decrease the chance
 // of any one controller predictably winning the race to schedule claims to a
@@ -48,45 +57,45 @@ const (
 // its competitors.
 type Jitterer func()
 
-// A ClaimSchedulingReconciler schedules resource claims to a resource class
-// that matches their class selector. Claims are reconciled by randomly
-// selecting a matching resource class and attempting to set it as the claim's
-// class reference. The ClaimSchedulingReconciler is designed for use in
-// claim scheduling controllers that race several others to schedule a claim.
-type ClaimSchedulingReconciler struct {
+// A Reconciler reconciles resource claims by setting their class reference to
+// the resource class annotated as the default. If multiple classes are
+// annotated as the default one of the annotated classes will be set, but which
+// one is undefined.
+type Reconciler struct {
 	client    client.Client
-	newClaim  func() Claim
-	classKind ClassKind
+	newClaim  func() resource.Claim
+	classKind resource.ClassKind
 	jitter    Jitterer
 }
 
-// A ClaimSchedulingReconcilerOption configures a ClaimSchedulingReconciler.
-type ClaimSchedulingReconcilerOption func(*ClaimSchedulingReconciler)
+// A ReconcilerOption configures a Reconciler.
+type ReconcilerOption func(*Reconciler)
 
-// WithSchedulingJitterer specifies the Jitterer a ClaimSchedulingReconciler should use.
-func WithSchedulingJitterer(j Jitterer) ClaimSchedulingReconcilerOption {
-	return func(r *ClaimSchedulingReconciler) {
+// WithDefaultingJitterer specifies the Jitterer a Reconciler should use.
+func WithDefaultingJitterer(j Jitterer) ReconcilerOption {
+	return func(r *Reconciler) {
 		r.jitter = j
 	}
 }
 
-// NewClaimSchedulingReconciler returns a ClaimSchedulingReconciler that
-// schedules resource claims to a resource class that matches their class
-// selector.
-func NewClaimSchedulingReconciler(m manager.Manager, of ClaimKind, to ClassKind, o ...ClaimSchedulingReconcilerOption) *ClaimSchedulingReconciler {
-	nc := func() Claim { return MustCreateObject(schema.GroupVersionKind(of), m.GetScheme()).(Claim) }
+// NewReconciler returns a Reconciler that sets the class reference of a
+// resource claim to the resource class annotated as the default.
+func NewReconciler(m manager.Manager, of resource.ClaimKind, to resource.ClassKind, o ...ReconcilerOption) *Reconciler {
+	nc := func() resource.Claim {
+		return resource.MustCreateObject(schema.GroupVersionKind(of), m.GetScheme()).(resource.Claim)
+	}
 
 	// Panic early if we've been asked to reconcile a claim or resource kind
 	// that has not been registered with our controller manager's scheme.
 	_ = nc()
 
-	r := &ClaimSchedulingReconciler{
+	r := &Reconciler{
 		client:    m.GetClient(),
 		newClaim:  nc,
 		classKind: to,
 		jitter: func() {
 			random := rand.New(rand.NewSource(time.Now().UnixNano()))
-			time.Sleep(time.Duration(random.Intn(claimSchedulingReconcileMaxJitterMs)) * time.Millisecond)
+			time.Sleep(time.Duration(random.Intn(claimDefaultingReconcileMaxJitterMs)) * time.Millisecond)
 		},
 	}
 
@@ -99,17 +108,17 @@ func NewClaimSchedulingReconciler(m manager.Manager, of ClaimKind, to ClassKind,
 
 // Reconcile a resource claim by using its class selector to select and allocate
 // it a resource class.
-func (r *ClaimSchedulingReconciler) Reconcile(req reconcile.Request) (reconcile.Result, error) {
-	log.V(logging.Debug).Info("Reconciling", "controller", claimSchedulingControllerName, "request", req)
+func (r *Reconciler) Reconcile(req reconcile.Request) (reconcile.Result, error) {
+	log.V(logging.Debug).Info("Reconciling", "controller", claimDefaultingControllerName, "request", req)
 
-	ctx, cancel := context.WithTimeout(context.Background(), claimSchedulingReconcileTimeout)
+	ctx, cancel := context.WithTimeout(context.Background(), claimDefaultingReconcileTimeout)
 	defer cancel()
 
 	claim := r.newClaim()
 	if err := r.client.Get(ctx, req.NamespacedName, claim); err != nil {
 		// There's no need to requeue if we no longer exist. Otherwise we'll be
 		// requeued implicitly because we return an error.
-		return reconcile.Result{}, errors.Wrap(IgnoreNotFound(err), errGetClaim)
+		return reconcile.Result{}, errors.Wrap(resource.IgnoreNotFound(err), errGetClaim)
 	}
 
 	// There could be several controllers racing to schedule this claim. If it
@@ -122,29 +131,37 @@ func (r *ClaimSchedulingReconciler) Reconcile(req reconcile.Request) (reconcile.
 	classes := &unstructured.UnstructuredList{}
 	classes.SetGroupVersionKind(r.classKind.List())
 
-	if err := r.client.List(ctx, classes, client.MatchingLabels(claim.GetClassSelector().MatchLabels)); err != nil {
-		// Claim scheduler controllers don't update the synced status because
+	if err := r.client.List(ctx, classes); err != nil {
+		// Claim defaulting controllers don't update the synced status because
 		// no one scheduler has the full view of whether the process failed or
 		// succeeded. It's possible another controller can successfully set a
 		// class even though we can't, so it would be confusing to mark this
 		// claim as failing to be reconciled. Instead we return an error - we'll
-		// be requeued but abort immediately if the claim was scheduled.
+		// be requeued but abort immediately if the claim was defaulted.
 		return reconcile.Result{}, errors.Wrap(err, errListClasses)
 	}
 
-	if len(classes.Items) == 0 {
-		// None of our classes matched the selector. We can't be sure whether
-		// another controller owns classes that matched the selector, or whether
-		// no classes match, so we requeue after a short wait. We'll abort the
-		// next reconcile immediately if another controller scheduled the claim.
+	defaults := []unstructured.Unstructured{}
+	for _, c := range classes.Items {
+		if c.GetAnnotations()[v1alpha1.AnnotationDefaultClassKey] == v1alpha1.AnnotationDefaultClassValue {
+			defaults = append(defaults, c)
+		}
+	}
+
+	if len(defaults) == 0 {
+		// None of our classes are annotated as the default. We can't be sure
+		// whether another controller owns the default class, or whether there
+		// is no default class, so we requeue after a short wait. We'll abort
+		// the next reconcile immediately if another controller defaulted the
+		// claim.
 		return reconcile.Result{RequeueAfter: aShortWait}, nil
 	}
 
 	random := rand.New(rand.NewSource(time.Now().UnixNano()))
-	selected := classes.Items[random.Intn(len(classes.Items))]
+	selected := defaults[random.Intn(len(defaults))]
 	claim.SetClassReference(meta.ReferenceTo(&selected, schema.GroupVersionKind(r.classKind)))
 
-	// There could be several controllers racing to schedule this claim to a
+	// There could be several controllers racing to default this claim to a
 	// class. We sleep for a randomly jittered amount of time before trying to
 	// update the class reference to decrease the chance of any one controller
 	// predictably winning the race, for example because it has fewer classes to
@@ -154,6 +171,6 @@ func (r *ClaimSchedulingReconciler) Reconcile(req reconcile.Request) (reconcile.
 	// Attempt to set the class reference. If a competing controller beat us
 	// we'll fail the write because the claim's resource version has changed
 	// since we read it. We'll be requeued, but will abort immediately if the
-	// claim was scheduled.
+	// claim was defaulted.
 	return reconcile.Result{Requeue: false}, errors.Wrap(r.client.Update(ctx, claim), errUpdateClaim)
 }
