+++ pkg/reconciler/claimdefaulting/reconciler_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resource
+package claimdefaulting
 
 import (
 	"strconv"
@@ -32,20 +32,22 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
+	"github.com/crossplaneio/crossplane-runtime/apis/core/v1alpha1"
+	"github.com/crossplaneio/crossplane-runtime/pkg/resource"
 	"github.com/crossplaneio/crossplane-runtime/pkg/resource/fake"
 	"github.com/crossplaneio/crossplane-runtime/pkg/test"
 )
 
-var _ reconcile.Reconciler = &ClaimDefaultingReconciler{}
+var _ reconcile.Reconciler = &Reconciler{}
 
-func TestClaimSchedulingReconciler(t *testing.T) {
+func TestReconciler(t *testing.T) {
 	name := "coolName"
 	uid := types.UID("definitely-a-uuid")
 
 	type args struct {
 		m  manager.Manager
-		of ClaimKind
-		to ClassKind
+		of resource.ClaimKind
+		to resource.ClassKind
 	}
 
 	type want struct {
@@ -65,8 +67,8 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 					Client: &test.MockClient{MockGet: test.NewMockGetFn(errBoom)},
 					Scheme: fake.SchemeWith(&fake.Claim{}),
 				},
-				of: ClaimKind(fake.GVK(&fake.Claim{})),
-				to: ClassKind(fake.GVK(&fake.Class{})),
+				of: resource.ClaimKind(fake.GVK(&fake.Claim{})),
+				to: resource.ClassKind(fake.GVK(&fake.Class{})),
 			},
 			want: want{err: errors.Wrap(errBoom, errGetClaim)},
 		},
@@ -76,8 +78,8 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 					Client: &test.MockClient{MockGet: test.NewMockGetFn(kerrors.NewNotFound(schema.GroupResource{}, ""))},
 					Scheme: fake.SchemeWith(&fake.Claim{}),
 				},
-				of: ClaimKind(fake.GVK(&fake.Claim{})),
-				to: ClassKind(fake.GVK(&fake.Class{})),
+				of: resource.ClaimKind(fake.GVK(&fake.Claim{})),
+				to: resource.ClassKind(fake.GVK(&fake.Class{})),
 			},
 			want: want{result: reconcile.Result{}},
 		},
@@ -93,8 +95,8 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 					},
 					Scheme: fake.SchemeWith(&fake.Claim{}),
 				},
-				of: ClaimKind(fake.GVK(&fake.Claim{})),
-				to: ClassKind(fake.GVK(&fake.Class{})),
+				of: resource.ClaimKind(fake.GVK(&fake.Claim{})),
+				to: resource.ClassKind(fake.GVK(&fake.Class{})),
 			},
 			want: want{result: reconcile.Result{Requeue: false}},
 		},
@@ -102,35 +104,35 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 			args: args{
 				m: &fake.Manager{
 					Client: &test.MockClient{
-						MockGet: test.NewMockGetFn(nil, func(o runtime.Object) error {
-							c := o.(*fake.Claim)
-							*c = fake.Claim{ClassSelector: fake.ClassSelector{Sel: &metav1.LabelSelector{}}}
-							return nil
-						}),
+						MockGet:  test.NewMockGetFn(nil),
 						MockList: test.NewMockListFn(errBoom),
 					},
 					Scheme: fake.SchemeWith(&fake.Claim{}),
 				},
-				of: ClaimKind(fake.GVK(&fake.Claim{})),
-				to: ClassKind(fake.GVK(&fake.Class{})),
+				of: resource.ClaimKind(fake.GVK(&fake.Claim{})),
+				to: resource.ClassKind(fake.GVK(&fake.Class{})),
 			},
 			want: want{err: errors.Wrap(errBoom, errListClasses)},
 		},
-		"NoClassesMatchLabels": {
+		"NoClassesAnnotatedDefault": {
 			args: args{
 				m: &fake.Manager{
 					Client: &test.MockClient{
-						MockGet: test.NewMockGetFn(nil, func(o runtime.Object) error {
-							c := o.(*fake.Claim)
-							*c = fake.Claim{ClassSelector: fake.ClassSelector{Sel: &metav1.LabelSelector{}}}
+						MockGet: test.NewMockGetFn(nil),
+						MockList: test.NewMockListFn(nil, func(o runtime.Object) error {
+							u := &unstructured.Unstructured{}
+							u.SetGroupVersionKind(fake.GVK(&fake.Class{}))
+							u.SetName(name)
+							u.SetUID(uid)
+							l := o.(*unstructured.UnstructuredList)
+							l.Items = []unstructured.Unstructured{*u}
 							return nil
 						}),
-						MockList: test.NewMockListFn(nil),
 					},
 					Scheme: fake.SchemeWith(&fake.Claim{}),
 				},
-				of: ClaimKind(fake.GVK(&fake.Claim{})),
-				to: ClassKind(fake.GVK(&fake.Class{})),
+				of: resource.ClaimKind(fake.GVK(&fake.Claim{})),
+				to: resource.ClassKind(fake.GVK(&fake.Class{})),
 			},
 			want: want{result: reconcile.Result{RequeueAfter: aShortWait}},
 		},
@@ -138,16 +140,13 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 			args: args{
 				m: &fake.Manager{
 					Client: &test.MockClient{
-						MockGet: test.NewMockGetFn(nil, func(o runtime.Object) error {
-							c := o.(*fake.Claim)
-							*c = fake.Claim{ClassSelector: fake.ClassSelector{Sel: &metav1.LabelSelector{}}}
-							return nil
-						}),
+						MockGet: test.NewMockGetFn(nil),
 						MockList: test.NewMockListFn(nil, func(o runtime.Object) error {
 							u := &unstructured.Unstructured{}
 							u.SetGroupVersionKind(fake.GVK(&fake.Class{}))
 							u.SetName(name)
 							u.SetUID(uid)
+							u.SetAnnotations(map[string]string{v1alpha1.AnnotationDefaultClassKey: v1alpha1.AnnotationDefaultClassValue})
 							l := o.(*unstructured.UnstructuredList)
 							l.Items = []unstructured.Unstructured{*u}
 							return nil
@@ -156,8 +155,8 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 					},
 					Scheme: fake.SchemeWith(&fake.Claim{}),
 				},
-				of: ClaimKind(fake.GVK(&fake.Claim{})),
-				to: ClassKind(fake.GVK(&fake.Class{})),
+				of: resource.ClaimKind(fake.GVK(&fake.Claim{})),
+				to: resource.ClassKind(fake.GVK(&fake.Class{})),
 			},
 			want: want{err: errors.Wrap(errBoom, errUpdateClaim)},
 		},
@@ -165,23 +164,19 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 			args: args{
 				m: &fake.Manager{
 					Client: &test.MockClient{
-						MockGet: test.NewMockGetFn(nil, func(o runtime.Object) error {
-							c := o.(*fake.Claim)
-							*c = fake.Claim{ClassSelector: fake.ClassSelector{Sel: &metav1.LabelSelector{}}}
-							return nil
-						}),
+						MockGet: test.NewMockGetFn(nil),
 						MockList: test.NewMockListFn(nil, func(o runtime.Object) error {
 							u := &unstructured.Unstructured{}
 							u.SetGroupVersionKind(fake.GVK(&fake.Class{}))
 							u.SetName(name)
 							u.SetUID(uid)
+							u.SetAnnotations(map[string]string{v1alpha1.AnnotationDefaultClassKey: v1alpha1.AnnotationDefaultClassValue})
 							l := o.(*unstructured.UnstructuredList)
 							l.Items = []unstructured.Unstructured{*u}
 							return nil
 						}),
 						MockUpdate: test.NewMockUpdateFn(nil, func(got runtime.Object) error {
 							want := &fake.Claim{}
-							want.SetClassSelector(&metav1.LabelSelector{})
 							want.SetClassReference(&corev1.ObjectReference{
 								APIVersion: fake.GVK(&fake.Class{}).GroupVersion().String(),
 								Kind:       fake.GVK(&fake.Class{}).Kind,
@@ -196,8 +191,8 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 					},
 					Scheme: fake.SchemeWith(&fake.Claim{}),
 				},
-				of: ClaimKind(fake.GVK(&fake.Claim{})),
-				to: ClassKind(fake.GVK(&fake.Class{})),
+				of: resource.ClaimKind(fake.GVK(&fake.Claim{})),
+				to: resource.ClassKind(fake.GVK(&fake.Class{})),
 			},
 			want: want{result: reconcile.Result{Requeue: false}},
 		},
@@ -205,7 +200,7 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			r := NewClaimSchedulingReconciler(tc.args.m, tc.args.of, tc.args.to, WithSchedulingJitterer(func() {}))
+			r := NewReconciler(tc.args.m, tc.args.of, tc.args.to, WithDefaultingJitterer(func() {}))
 			got, err := r.Reconcile(reconcile.Request{})
 
 			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
@@ -219,7 +214,7 @@ func TestClaimSchedulingReconciler(t *testing.T) {
 	}
 }
 
-func TestClaimSchedulingReconcilerRandomness(t *testing.T) {
+func TestReconcilerRandomness(t *testing.T) {
 	classes := 10
 	reconciles := 100
 	refs := make([]*corev1.ObjectReference, 0)
@@ -227,6 +222,7 @@ func TestClaimSchedulingReconcilerRandomness(t *testing.T) {
 	newClass := func(i int) unstructured.Unstructured {
 		u := &unstructured.Unstructured{}
 		u.SetUID(types.UID(strconv.Itoa(i)))
+		u.SetAnnotations(map[string]string{v1alpha1.AnnotationDefaultClassKey: v1alpha1.AnnotationDefaultClassValue})
 		return *u
 	}
 
@@ -245,7 +241,7 @@ func TestClaimSchedulingReconcilerRandomness(t *testing.T) {
 				return nil
 			}),
 			MockUpdate: test.NewMockUpdateFn(nil, func(obj runtime.Object) error {
-				ls := obj.(ClassReferencer)
+				ls := obj.(resource.ClassReferencer)
 				refs = append(refs, ls.GetClassReference())
 				return nil
 			}),
@@ -253,10 +249,10 @@ func TestClaimSchedulingReconcilerRandomness(t *testing.T) {
 		Scheme: fake.SchemeWith(&fake.Claim{}),
 	}
 
-	r := NewClaimSchedulingReconciler(m,
-		ClaimKind(fake.GVK(&fake.Claim{})),
-		ClassKind(fake.GVK(&fake.Class{})),
-		WithSchedulingJitterer(func() {}))
+	r := NewReconciler(m,
+		resource.ClaimKind(fake.GVK(&fake.Claim{})),
+		resource.ClassKind(fake.GVK(&fake.Class{})),
+		WithDefaultingJitterer(func() {}))
 
 	for i := 0; i < reconciles; i++ {
 		r.Reconcile(reconcile.Request{})
