+++ pkg/reconciler/managed/api.go
@@ -1,0 +1,138 @@
+/*
+Copyright 2019 The Crossplane Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package resource
+
+import (
+	"context"
+
+	"github.com/pkg/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	util "sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+
+	"github.com/crossplaneio/crossplane-runtime/pkg/meta"
+	"github.com/crossplaneio/crossplane-runtime/pkg/resource"
+)
+
+// Error strings.
+const (
+	errSecretConflict       = "cannot establish control of existing connection secret"
+	errCreateOrUpdateSecret = "cannot create or update connection secret"
+	errUpdateManaged        = "cannot update managed resource"
+	errUpdateManagedStatus  = "cannot update managed resource status"
+)
+
+// An APIManagedFinalizer adds and removes finalizers to and from a resource.
+type APIManagedFinalizer struct {
+	client    client.Client
+	finalizer string
+}
+
+// NewAPIManagedFinalizer returns a new APIManagedFinalizer.
+func NewAPIManagedFinalizer(c client.Client, finalizer string) *APIManagedFinalizer {
+	return &APIManagedFinalizer{client: c, finalizer: finalizer}
+}
+
+// AddFinalizer to the supplied Managed resource.
+func (a *APIManagedFinalizer) AddFinalizer(ctx context.Context, mg resource.Managed) error {
+	if meta.FinalizerExists(mg, a.finalizer) {
+		return nil
+	}
+	meta.AddFinalizer(mg, a.finalizer)
+	return errors.Wrap(a.client.Update(ctx, mg), errUpdateManaged)
+}
+
+// RemoveFinalizer from the supplied Managed resource.
+func (a *APIManagedFinalizer) RemoveFinalizer(ctx context.Context, mg resource.Managed) error {
+	meta.RemoveFinalizer(mg, a.finalizer)
+	return errors.Wrap(resource.IgnoreNotFound(a.client.Update(ctx, mg)), errUpdateManaged)
+}
+
+// ManagedNameAsExternalName writes the name of the managed resource to
+// the external name annotation field in order to be used as name of
+// the external resource in provider.
+type ManagedNameAsExternalName struct{ client client.Client }
+
+// NewManagedNameAsExternalName returns a new ManagedNameAsExternalName.
+func NewManagedNameAsExternalName(c client.Client) *ManagedNameAsExternalName {
+	return &ManagedNameAsExternalName{client: c}
+}
+
+// Initialize the given managed resource.
+func (a *ManagedNameAsExternalName) Initialize(ctx context.Context, mg resource.Managed) error {
+	if meta.GetExternalName(mg) != "" {
+		return nil
+	}
+	meta.SetExternalName(mg, mg.GetName())
+	return errors.Wrap(a.client.Update(ctx, mg), errUpdateManaged)
+}
+
+// An APISecretPublisher publishes ConnectionDetails by submitting a Secret to a
+// Kubernetes API server.
+type APISecretPublisher struct {
+	client client.Client
+	typer  runtime.ObjectTyper
+}
+
+// NewAPISecretPublisher returns a new APISecretPublisher.
+func NewAPISecretPublisher(c client.Client, ot runtime.ObjectTyper) *APISecretPublisher {
+	return &APISecretPublisher{client: c, typer: ot}
+}
+
+// PublishConnection publishes the supplied ConnectionDetails to a Secret in the
+// same namespace as the supplied Managed resource. It is a no-op if the secret
+// already exists with the supplied ConnectionDetails.
+func (a *APISecretPublisher) PublishConnection(ctx context.Context, mg resource.Managed, c ConnectionDetails) error {
+	// This resource does not want to expose a connection secret.
+	if mg.GetWriteConnectionSecretToReference() == nil {
+		return nil
+	}
+
+	s := resource.ConnectionSecretFor(mg, resource.MustGetKind(mg, a.typer))
+
+	_, err := util.CreateOrUpdate(ctx, a.client, s, func() error {
+		// Inside this anonymous function s could either be unchanged (if it
+		// does not exist in the API server) or updated to reflect its current
+		// state according to the API server.
+		if c := metav1.GetControllerOf(s); c == nil || c.UID != mg.GetUID() {
+			return errors.New(errSecretConflict)
+		}
+
+		// NOTE(negz): We want to support additive publishing, i.e. support
+		// setting one subset of secret values then later setting another subset
+		// without effecting the original subset.
+		if s.Data == nil {
+			s.Data = make(map[string][]byte, len(c))
+		}
+
+		for k, v := range c {
+			s.Data[k] = v
+		}
+
+		return nil
+	})
+
+	return errors.Wrap(err, errCreateOrUpdateSecret)
+}
+
+// UnpublishConnection is no-op since PublishConnection only creates resources
+// that will be garbage collected by Kubernetes when the managed resource is
+// deleted.
+func (a *APISecretPublisher) UnpublishConnection(ctx context.Context, mg resource.Managed, c ConnectionDetails) error {
+	return nil
+}
