+++ pkg/reconciler/managed/api_test.go
@@ -31,81 +31,207 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/crossplaneio/crossplane-runtime/apis/core/v1alpha1"
+	"github.com/crossplaneio/crossplane-runtime/pkg/meta"
+	"github.com/crossplaneio/crossplane-runtime/pkg/resource"
 	"github.com/crossplaneio/crossplane-runtime/pkg/resource/fake"
 	"github.com/crossplaneio/crossplane-runtime/pkg/test"
 )
 
 var (
-	_ ManagedConnectionPublisher = &APISecretPublisher{}
-	_ ManagedConnectionPublisher = PublisherChain{}
+	_ ManagedFinalizer   = &APIManagedFinalizer{}
+	_ ManagedInitializer = &ManagedNameAsExternalName{}
 )
 
-func TestPublisherChain(t *testing.T) {
+func TestManagedRemoveFinalizer(t *testing.T) {
+	finalizer := "veryfinal"
+
 	type args struct {
 		ctx context.Context
-		mg  Managed
-		c   ConnectionDetails
+		mg  resource.Managed
+	}
+
+	type want struct {
+		err error
+		mg  resource.Managed
 	}
 
 	errBoom := errors.New("boom")
 
 	cases := map[string]struct {
-		p    ManagedConnectionPublisher
-		args args
-		want error
+		client client.Client
+		args   args
+		want   want
 	}{
-		"EmptyChain": {
-			p: PublisherChain{},
+		"UpdateManagedError": {
+			client: &test.MockClient{MockUpdate: test.NewMockUpdateFn(errBoom)},
 			args: args{
 				ctx: context.Background(),
-				mg:  &fake.Managed{},
-				c:   ConnectionDetails{},
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Finalizers: []string{finalizer}}},
 			},
-			want: nil,
+			want: want{
+				err: errors.Wrap(errBoom, errUpdateManaged),
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Finalizers: []string{}}},
+			},
 		},
-		"SuccessfulPublisher": {
-			p: PublisherChain{
-				ManagedConnectionPublisherFns{
-					PublishConnectionFn: func(_ context.Context, mg Managed, c ConnectionDetails) error {
-						return nil
-					},
-					UnpublishConnectionFn: func(ctx context.Context, mg Managed, c ConnectionDetails) error {
-						return nil
-					},
-				},
-			},
+		"Successful": {
+			client: &test.MockClient{MockUpdate: test.NewMockUpdateFn(nil)},
 			args: args{
 				ctx: context.Background(),
-				mg:  &fake.Managed{},
-				c:   ConnectionDetails{},
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Finalizers: []string{finalizer}}},
 			},
-			want: nil,
-		},
-		"PublisherReturnsError": {
-			p: PublisherChain{
-				ManagedConnectionPublisherFns{
-					PublishConnectionFn: func(_ context.Context, mg Managed, c ConnectionDetails) error {
-						return errBoom
-					},
-					UnpublishConnectionFn: func(ctx context.Context, mg Managed, c ConnectionDetails) error {
-						return nil
-					},
-				},
+			want: want{
+				err: nil,
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Finalizers: []string{}}},
 			},
-			args: args{
-				ctx: context.Background(),
-				mg:  &fake.Managed{},
-				c:   ConnectionDetails{},
-			},
-			want: errBoom,
 		},
 	}
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			got := tc.p.PublishConnection(tc.args.ctx, tc.args.mg, tc.args.c)
-			if diff := cmp.Diff(tc.want, got, test.EquateErrors()); diff != "" {
-				t.Errorf("Publish(...): -want, +got:\n%s", diff)
+			api := NewAPIManagedFinalizer(tc.client, finalizer)
+			err := api.RemoveFinalizer(tc.args.ctx, tc.args.mg)
+			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
+				t.Errorf("api.RemoveFinalizer(...): -want error, +got error:\n%s", diff)
+			}
+			if diff := cmp.Diff(tc.want.mg, tc.args.mg, test.EquateConditions()); diff != "" {
+				t.Errorf("api.RemoveFinalizer(...) Managed: -want, +got:\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestAPIManagedFinalizerAdder(t *testing.T) {
+	finalizer := "veryfinal"
+
+	type args struct {
+		ctx context.Context
+		mg  resource.Managed
+	}
+
+	type want struct {
+		err error
+		mg  resource.Managed
+	}
+
+	errBoom := errors.New("boom")
+
+	cases := map[string]struct {
+		client client.Client
+		args   args
+		want   want
+	}{
+		"UpdateManagedError": {
+			client: &test.MockClient{MockUpdate: test.NewMockUpdateFn(errBoom)},
+			args: args{
+				ctx: context.Background(),
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Finalizers: []string{}}},
+			},
+			want: want{
+				err: errors.Wrap(errBoom, errUpdateManaged),
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Finalizers: []string{finalizer}}},
+			},
+		},
+		"Successful": {
+			client: &test.MockClient{MockUpdate: test.NewMockUpdateFn(nil)},
+			args: args{
+				ctx: context.Background(),
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Finalizers: []string{}}},
+			},
+			want: want{
+				err: nil,
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Finalizers: []string{finalizer}}},
+			},
+		},
+	}
+
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			api := NewAPIManagedFinalizer(tc.client, finalizer)
+			err := api.AddFinalizer(tc.args.ctx, tc.args.mg)
+			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
+				t.Errorf("api.Initialize(...): -want error, +got error:\n%s", diff)
+			}
+			if diff := cmp.Diff(tc.want.mg, tc.args.mg, test.EquateConditions()); diff != "" {
+				t.Errorf("api.Initialize(...) Managed: -want, +got:\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestManagedNameAsExternalName(t *testing.T) {
+	type args struct {
+		ctx context.Context
+		mg  resource.Managed
+	}
+
+	type want struct {
+		err error
+		mg  resource.Managed
+	}
+
+	errBoom := errors.New("boom")
+	testExternalName := "my-external-name"
+
+	cases := map[string]struct {
+		client client.Client
+		args   args
+		want   want
+	}{
+		"UpdateManagedError": {
+			client: &test.MockClient{MockUpdate: test.NewMockUpdateFn(errBoom)},
+			args: args{
+				ctx: context.Background(),
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Name: testExternalName}},
+			},
+			want: want{
+				err: errors.Wrap(errBoom, errUpdateManaged),
+				mg: &fake.Managed{ObjectMeta: metav1.ObjectMeta{
+					Name:        testExternalName,
+					Annotations: map[string]string{meta.ExternalNameAnnotationKey: testExternalName},
+				}},
+			},
+		},
+		"UpdateSuccessful": {
+			client: &test.MockClient{MockUpdate: test.NewMockUpdateFn(nil)},
+			args: args{
+				ctx: context.Background(),
+				mg:  &fake.Managed{ObjectMeta: metav1.ObjectMeta{Name: testExternalName}},
+			},
+			want: want{
+				err: nil,
+				mg: &fake.Managed{ObjectMeta: metav1.ObjectMeta{
+					Name:        testExternalName,
+					Annotations: map[string]string{meta.ExternalNameAnnotationKey: testExternalName},
+				}},
+			},
+		},
+		"UpdateNotNeeded": {
+			args: args{
+				ctx: context.Background(),
+				mg: &fake.Managed{ObjectMeta: metav1.ObjectMeta{
+					Name:        testExternalName,
+					Annotations: map[string]string{meta.ExternalNameAnnotationKey: "some-name"},
+				}},
+			},
+			want: want{
+				err: nil,
+				mg: &fake.Managed{ObjectMeta: metav1.ObjectMeta{
+					Name:        testExternalName,
+					Annotations: map[string]string{meta.ExternalNameAnnotationKey: "some-name"},
+				}},
+			},
+		},
+	}
+
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			api := NewManagedNameAsExternalName(tc.client)
+			err := api.Initialize(tc.args.ctx, tc.args.mg)
+			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
+				t.Errorf("api.Initialize(...): -want error, +got error:\n%s", diff)
+			}
+			if diff := cmp.Diff(tc.want.mg, tc.args.mg, test.EquateConditions()); diff != "" {
+				t.Errorf("api.Initialize(...) Managed: -want, +got:\n%s", diff)
 			}
 		})
 	}
@@ -119,7 +245,7 @@ func TestAPISecretPublisher(t *testing.T) {
 
 	type args struct {
 		ctx context.Context
-		mg  Managed
+		mg  resource.Managed
 		c   ConnectionDetails
 	}
